requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
openai>=1.0.0
imageio-ffmpeg>=0.5.1
